+++ packages/survey-creator-core/tests/creator-base.tests.ts
@@ -61,6 +61,7 @@ test("Select new added question", (): any => {
     elements: [{ type: "text", name: "question1" }]
   };
   expect(creator.activeTab).toEqual("designer");
+  // eslint-disable-next-line no-self-assign
   creator.survey.currentPage = creator.survey.currentPage;
   creator.clickToolboxItem({ type: "text" });
   expect(creator.selectedElementName).toEqual("question2");
@@ -100,6 +101,7 @@ test("Update JSON before drag&drop", (): any => {
     elements: [{ type: "text", name: "question1" }]
   };
   expect(creator.activeTab).toEqual("designer");
+  // eslint-disable-next-line no-self-assign
   creator.survey.currentPage = creator.survey.currentPage;
   var json: any = {
     type: "panel",
@@ -896,7 +898,7 @@ test("Test plug-ins JSON-Text in creator, autosave", (): any => {
   const json = {
     pages: [
       {
-        elements: [{type: "text", name: "q1"}]
+        elements: [{ type: "text", name: "q1" }]
       }
     ]
   };
@@ -1240,4 +1242,34 @@ test("Set readOnly option", (): any => {
   } catch (e) {
     expect(e).toBeNull();
   }
+});
+test("Set allowEditSurveyTitle option", (): any => {
+  const creator = new CreatorTester({ allowEditSurveyTitle: false });
+  expect(creator.allowEditSurveyTitle).toBeFalsy();
+  expect(Serializer.findProperty("survey", "title").visible).toBeFalsy();
+  creator.allowEditSurveyTitle = true;
+  expect(Serializer.findProperty("survey", "title").visible).toBeTruthy();
+});
+test("creator.onActiveTabChanged", (): any => {
+  const creator = new CreatorTester({
+    showTranslationTab: true,
+    showLogicTab: true,
+  });
+  let tabName;
+  let plugin;
+  let model;
+  creator.onActiveTabChanged.add((sender, options) => {
+    tabName = options.tabName;
+    plugin = options.plugin;
+    model = options.model;
+  });
+  expect(creator.viewType).toEqual("designer");
+  creator.makeNewViewActive("test");
+  expect(tabName).toEqual("test");
+  expect(plugin).toEqual(creator.getPlugin("test"));
+  expect(model).toEqual(plugin.model);
+  creator.makeNewViewActive("logic");
+  expect(tabName).toEqual("logic");
+  expect(plugin).toEqual(creator.getPlugin("logic"));
+  expect(model).toEqual(plugin.model);
 });