+++ packages/survey-creator/src/questionEditors/questionEditorDefinition.ts
@@ -1,5 +1,3 @@
-import * as Survey from "survey-knockout";
-
 export interface ISurveyQuestionEditorDefinition {
   title?: string;
   properties?: Array<string | { name: string, title?: string, tab?: string }>;
@@ -59,6 +57,8 @@ export class SurveyQuestionEditorDefinition {
         "placeHolder",
         { name: "maxLength", tab: "validation" },
         { name: "textUpdateMode", tab: "data" },
+        { name: "autoGrow", tab: "layout" },
+        { name: "multiLine", tab: "layout" },
       ],
     },
     file: {
@@ -74,6 +74,7 @@ export class SurveyQuestionEditorDefinition {
         "imageHeight",
         "imageWidth",
         "waitForUpload",
+        "needConfirmRemoveFile"
       ],
     },
     html: {
@@ -102,6 +103,7 @@ export class SurveyQuestionEditorDefinition {
       properties: [
         "allowAddRows",
         "allowRemoveRows",
+        "allowRowsDragAndDrop",
         "rowCount",
         "minRowCount",
         "maxRowCount",
@@ -131,6 +133,7 @@ export class SurveyQuestionEditorDefinition {
         { name: "isAllRowRequired", tab: "validation" },
         { name: "showHeader", tab: "layout" },
         { name: "rowsOrder", tab: "rows" },
+        { name: "hideIfRowsEmpty", tab: "rows" },
         { name: "columnsVisibleIf", tab: "logic" },
         { name: "rowsVisibleIf", tab: "logic" },
       ],
@@ -195,6 +198,7 @@ export class SurveyQuestionEditorDefinition {
     dropdown: {
       properties: [
         "optionsCaption",
+        "autoComplete",
         { name: "choicesMin", tab: "choices" },
         { name: "choicesMax", tab: "choices" },
         { name: "choicesStep", tab: "choices" },
@@ -218,6 +222,7 @@ export class SurveyQuestionEditorDefinition {
         "imageFit",
         "imageHeight",
         "imageWidth",
+        "text"
       ],
     },
     "itemvalue[]@choices": {
@@ -517,6 +522,7 @@ export class SurveyQuestionEditorDefinition {
         { name: "questionsOrder", tab: "question" },
         { name: "maxTextLength", tab: "question" },
         { name: "maxOthersLength", tab: "question" },
+        { name: "autoGrowComment", tab: "question" },
 
         { name: "showPageTitles", tab: "pages" },
         { name: "showPageNumbers", tab: "pages" },